backend/requirements.txt
fastapi
uvicorn
motor
pydantic
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
python-multipart